+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Backwards incompatible changes for 0.3.0-rc14
  * Changed release adapter to be configurable (adapter and bucket), check new variables to be set in the docs.
+ * Created secrets adapter (adapter and path) to be able to configure different formats of fetching secrets.
 
 ### Bug fixes
  * [[`PR-51`](https://github.com/thiagoesteves/deployex/pull/51)] Terminal copy/paste bug, terminal was pasting when copying code within it.