+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.3.0-rc15 ()
 
 ## Backwards incompatible changes for 0.3.0-rc14
- * Changed release adapter to be configurable (adapter and bucket), check new variables to be set in the docs.
- * Created secrets adapter (adapter and path) to be able to configure different formats of fetching secrets.
+ * Changed release adapter to be configurable (adapter and bucket), new variables documentation added to the README.
+ * Created secrets adapter (adapter and path) to be able to configure different formats of fetching secrets, new variables documentation added to the README.
+ * Added new configuration for google credentials when using GCP, new variables documentation added to the README.
 
 ### Bug fixes
  * [[`PR-51`](https://github.com/thiagoesteves/deployex/pull/51)] Terminal copy/paste bug, terminal was pasting when copying code within it.